convert Auth/wrapper to typescript and resolve type issues
